@@ -9,22 +9,24 @@
     "type": "git",
     "url": "https://github.com/storybookjs/storybook-addon-kit"
   },
+  "type": "module",
   "license": "MIT",
   "author": "package-author",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "node": "./dist/index.js",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js"
     },
-    "./manager": "./dist/manager.mjs",
-    "./preview": "./dist/preview.mjs",
+    "./preview": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/preview.js",
+      "require": "./dist/preview.js"
+    },
+    "./preset": "./dist/preset.cjs",
+    "./manager": "./dist/manager.js",
     "./package.json": "./package.json"
   },
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
   "files": [
     "dist/**/*",
     "README.md",
@@ -39,54 +41,63 @@
     "build:watch": "npm run build -- --watch",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "run-p build:watch 'storybook --quiet'",
-    "prerelease": "zx scripts/prepublish-checks.mjs",
+    "prerelease": "zx scripts/prepublish-checks.js",
     "release": "npm run build && auto shipit",
-    "eject-ts": "zx scripts/eject-typescript.mjs",
+    "eject-ts": "zx scripts/eject-typescript.js",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "7.0.0",
-    "@storybook/addon-interactions": "7.0.0",
-    "@storybook/addon-links": "7.0.0",
-    "@storybook/blocks": "7.0.0",
-    "@storybook/components": "7.0.0",
-    "@storybook/core-events": "7.0.0",
-    "@storybook/manager": "^7.6.0",
-    "@storybook/manager-api": "7.0.0",
-    "@storybook/preview": "^7.6.0",
-    "@storybook/preview-api": "7.0.0",
-    "@storybook/react": "7.0.0",
-    "@storybook/react-vite": "7.0.0",
+    "@storybook/addon-essentials": "8.0.0-beta.2",
+    "@storybook/addon-interactions": "8.0.0-beta.2",
+    "@storybook/addon-links": "8.0.0-beta.2",
+    "@storybook/blocks": "8.0.0-beta.2",
+    "@storybook/components": "8.0.0-beta.2",
+    "@storybook/core-events": "8.0.0-beta.2",
+    "@storybook/manager": "8.0.0-beta.2",
+    "@storybook/manager-api": "8.0.0-beta.2",
+    "@storybook/preview": "8.0.0-beta.2",
+    "@storybook/preview-api": "8.0.0-beta.2",
+    "@storybook/react": "8.0.0-beta.2",
+    "@storybook/react-vite": "8.0.0-beta.2",
     "@storybook/testing-library": "^0.2.2",
-    "@storybook/theming": "7.0.0",
-    "@storybook/types": "7.0.0",
+    "@storybook/theming": "8.0.0-beta.2",
+    "@storybook/types": "8.0.0-beta.2",
     "@types/node": "^18.15.0",
-    "@types/react": "^16.14.51",
-    "@types/react-dom": "^16.9.22",
-    "@vitejs/plugin-react": "^4.1.1",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
+    "@vitejs/plugin-react": "^4.2.1",
     "auto": "^11.0.4",
     "boxen": "^7.1.1",
     "dedent": "^1.5.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^3.0.3",
+    "prettier": "^3.2.5",
     "prompts": "^2.4.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^5.0.5",
-    "storybook": "7.0.0",
-    "tsup": "^7.2.0",
-    "typescript": "^5.2.2",
-    "vite": "^4.5.0",
+    "storybook": "8.0.0-beta.2",
+    "tsup": "^8.0.2",
+    "typescript": "^5.3.3",
+    "vite": "^5.1.2",
     "zx": "^7.2.3"
   },
   "publishConfig": {
     "access": "public"
   },
   "bundler": {
-    "exportEntries": ["src/index.ts"],
-    "managerEntries": ["src/manager.ts"],
-    "previewEntries": ["src/preview.ts"]
+    "exportEntries": [
+      "src/index.ts"
+    ],
+    "managerEntries": [
+      "src/manager.ts"
+    ],
+    "previewEntries": [
+      "src/preview.ts"
+    ],
+    "nodeEntries": [
+      "src/preset.ts"
+    ]
   },
   "storybook": {
     "displayName": "Addon Kit",
